+++ py_trees/demos/pick_up_where_you_left_off.py
@@ -1,0 +1,173 @@
+#!/usr/bin/env python
+#
+# License: BSD
+#   https://raw.githubusercontent.com/stonier/py_trees/devel/LICENSE
+#
+##############################################################################
+# Documentation
+##############################################################################
+
+"""
+.. argparse::
+   :module: py_trees.demos.pick_up_where_you_left_off
+   :func: command_line_argument_parser
+   :prog: py-trees-demo-pick-up-where-you-left-off
+
+.. graphviz:: dot/pick_up_where_you_left_off.dot
+
+.. image:: images/pick_up_where_you_left_off.gif
+"""
+
+##############################################################################
+# Imports
+##############################################################################
+
+import argparse
+import functools
+import py_trees
+import sys
+import time
+
+import py_trees.console as console
+
+##############################################################################
+# Classes
+##############################################################################
+
+
+def description(root):
+    content = "A demonstration of the 'pick up where you left off' idiom.\n\n"
+    content += "A common behaviour tree pattern that allows you to resume\n"
+    content += "work after being interrupted by a high priority interrupt.\n"
+    content += "\n"
+    content += "EVENTS\n"
+    content += "\n"
+    content += " -  2 : task one done, task two running\n"
+    content += " -  3 : high priority interrupt\n"
+    content += " -  7 : task two restarts\n"
+    content += " -  9 : task two done\n"
+    content += "\n"
+    if py_trees.console.has_colours:
+        banner_line = console.green + "*" * 79 + "\n" + console.reset
+        s = "\n"
+        s += banner_line
+        s += console.bold_white + "Trees".center(79) + "\n" + console.reset
+        s += banner_line
+        s += "\n"
+        s += content
+        s += "\n"
+        s += banner_line
+    else:
+        s = content
+    return s
+
+
+def epilog():
+    if py_trees.console.has_colours:
+        return console.cyan + "And his noodly appendage reached forth to tickle the blessed...\n" + console.reset
+    else:
+        return None
+
+
+def command_line_argument_parser():
+    parser = argparse.ArgumentParser(description=description(create_tree()),
+                                     epilog=epilog(),
+                                     formatter_class=argparse.RawDescriptionHelpFormatter,
+                                     )
+    group = parser.add_mutually_exclusive_group()
+    group.add_argument('-r', '--render', action='store_true', help='render dot tree to file')
+    group.add_argument('-i', '--interactive', action='store_true', help='pause and wait for keypress at each tick')
+    return parser
+
+
+def pre_tick_handler(behaviour_tree):
+    """
+    This prints a banner and will run immediately before every tick of the tree.
+
+    Args:
+        behaviour_tree (:class:`~py_trees.trees.BehaviourTree`): the tree custodian
+
+    """
+    print("\n--------- Run %s ---------\n" % behaviour_tree.count)
+
+
+def post_tick_handler(snapshot_visitor, behaviour_tree):
+    """
+    Prints an ascii tree with the current snapshot status.
+    """
+    print("\n" + py_trees.display.ascii_tree(behaviour_tree.root,
+                                             snapshot_information=snapshot_visitor))
+
+def create_tree():
+    task_one = py_trees.behaviours.Count(
+        name="Task 1",
+        fail_until=0,
+        running_until=2,
+        success_until=10
+    )
+    task_two = py_trees.behaviours.Count(
+        name="Task 2",
+        fail_until=0,
+        running_until=2,
+        success_until=10
+    )
+    high_priority_interrupt = py_trees.meta.running_is_failure(
+        py_trees.behaviours.Periodic)(
+        name="High Priority",
+        n=3
+    )
+    piwylo = py_trees.idioms.pick_up_where_you_left_off(
+        name="Pick Up\nWhere You\nLeft Off",
+        tasks=[task_one, task_two]
+    )
+    root = py_trees.composites.Selector(name="Root")
+    root.add_children([high_priority_interrupt, piwylo])
+
+    return root
+
+##############################################################################
+# Main
+##############################################################################
+
+def main():
+    """
+    Entry point for the demo script.
+    """
+    args = command_line_argument_parser().parse_args()
+    py_trees.logging.level = py_trees.logging.Level.DEBUG
+    tree = create_tree()
+    print(description(tree))
+
+    ####################
+    # Rendering
+    ####################
+    if args.render:
+        py_trees.display.render_dot_tree(tree)
+        sys.exit()
+
+    ####################
+    # Tree Stewardship
+    ####################
+    behaviour_tree = py_trees.trees.BehaviourTree(tree)
+    behaviour_tree.add_pre_tick_handler(pre_tick_handler)
+    behaviour_tree.visitors.append(py_trees.visitors.DebugVisitor())
+    snapshot_visitor = py_trees.visitors.SnapshotVisitor()
+    behaviour_tree.add_post_tick_handler(functools.partial(post_tick_handler, snapshot_visitor))
+    behaviour_tree.visitors.append(snapshot_visitor)
+    behaviour_tree.setup(timeout=15)
+
+    ####################
+    # Tick Tock
+    ####################
+    if args.interactive:
+        py_trees.console.read_single_keypress()
+    while True:
+        try:
+            behaviour_tree.tick()
+            if args.interactive:
+                py_trees.console.read_single_keypress()
+            else:
+                time.sleep(0.5)
+        except KeyboardInterrupt:
+            break
+    print("\n")